atualização do repositório com novos arquivos

+++ Identificacao_de_Sistemas/firmwares_microcontroladores/PlatformIo/Esp32_ttgo_modulos/lib/referencia/src/referencia.h
@@ -53,8 +53,8 @@ public:
         ampl = ampl_;
         offset = offset_;
         Ts = Ts_;
-        Tsf = 1.0 / freq;
-        Tsf_init = 1.0 / freq;
+        Tsf = (float)1.0 / freq;
+        Tsf_init = (float)1.0 / freq;
         sinal = 0.0;
     }
